feat(network): implement arena for any actions that is somehow affects the server data
CarryArena is required for an action that performs put or post, or similar request and changes the server data. We create a separate interface for that because default Arena already have a method with the "fetch" occurrence. Accordingly, we might require an additional class for result from this arena, because if request failed it will be simplifier to implement a retrying using this class.