feat(application, android, screen, comment, dispatch): implement bottom panel with only send icon
Bottom panel is require for text formatting while typing a message. It is not implemented right now, but the space is already taken and todo message displayed. This panel placed at the bottom of the screen, above the bottom sheet, so the panel will be always visible and able to use. The content, where user types a message also is scrollable and scroll events from bottom sheet and scrollable EditText will not being in the conflict.